+++ ruby/test-integration/replication_agent_spec.rb
@@ -69,4 +69,14 @@ describe 'Replication Agent' do
     )
     expect(status_code).to eq(200)
   end
+
+  describe 'test agents list' do
+
+    it 'should succeed' do
+      data, status_code, headers = @sling.get_agents_with_http_info('author')
+      expect(status_code).to eq(200)
+    end
+
+  end
+
 end