+++ ruby/CHANGELOG.md
@@ -1,3 +1,6 @@
+### 0.9.7
+* Add sling get agents support #16
+
 ### 0.9.6
 * Add Apache Sling DavEx Servlet OSGI config support #11
 * Add Apache Sling Referrer Filter OSGI config support